+++ changelog.md
@@ -6,6 +6,7 @@ Codion Change Log
 - FilePreferencesFactory removed, UserPreferences.file() added.
 ### is.codion.swing.common.ui
 - FilterTableCellRenderer.Builder.ColumnClassBuilder added.
+- FilterTable.Builder.rowHeight(), rowMargin(), intercellSpacing(), gridColor(), showGrid(), showHorizontalLines(), showVerticalLines() and dragEnabled() added.
 
 ## 0.18.40
 ### is.codion